team features

+++ server/lib/game_machine/game_systems/team_manager.rb
@@ -4,8 +4,13 @@ module GameMachine
       include Models
       include Commands
 
+      attr_reader :team_handler
       def post_init(*args)
+        if Application.config.team_handler
+          @team_handler = Application.config.team_handler.constantize.new
+        end
         define_update_procs
+        commands.misc.client_manager_register(self.class.name)
       end
 
       def chat_channel(topic,flags)
@@ -45,6 +50,21 @@ module GameMachine
         player_team.save!
       end
 
+      def send_team(message)
+        if player_team = PlayerTeam.find!(message.player_id)
+          if team = Team.find!(player_team.name)
+            commands.player.send_message(team,message.player_id)
+          end
+        end
+      end
+
+      def teams_request(message)
+        if teams = Teams.find('teams')
+          commands.player.send_message(teams,message.player_id)
+        end
+        send_team(message)
+      end
+
       def destroy_player_team(player_id)
         if player_team = PlayerTeam.find!(player_id)
           player_team.destroy!
@@ -55,11 +75,46 @@ module GameMachine
         Uniqueid.generate_token(team_name)
       end
 
+      def default_max_members
+        100
+      end
+
+      def can_add_member?(team_name,player_id)
+        return true unless team_handler
+        team_handler.can_add_member?(team_name, player_id)
+      end
+
+      def can_create_team?(team_name,player_id)
+        return true unless team_handler
+        team_handler.can_create_team?(team_name, player_id)
+      end
+
+      def handler_team_added(team_name)
+        return true unless team_handler
+        team_handler.team_added(team_name)
+      end
+
+      def handler_team_removed(team_name)
+        return true unless team_handler
+        team_handler.team_added(team_name)
+      end
+
+      def destroy_on_owner_leave?
+        return true unless team_handler
+        team_handler.destroy_on_owner_leave?
+      end
+
       def create_team(message)
+        unless can_create_team?(message.name,message.player_id)
+          return false
+        end
+
         if team = Team.find!(message.name)
           send_team_joined(team.name,message.player_id)
           return false
         end
+
+        message.max_members ||= default_max_members
 
         team = Team.new(
           :id => message.name,
@@ -67,6 +122,8 @@ module GameMachine
           :name => message.name,
           :owner => message.owner,
           :access => message.access,
+          :max_members => message.max_members,
+          :destroy_on_owner_leave => destroy_on_owner_leave?,
           :members => [message.owner]
         )
 
@@ -79,12 +136,22 @@ module GameMachine
         commands.datastore.call_dbproc(:team_add_team,'teams',team.to_storage_entity,false)
         join_chat(team.name,team.owner)
         send_team_joined(team.name,message.player_id)
-        GameMachine.logger.info "Team #{team.name} created"
+        handler_team_added(team.name)
       end
 
-      def destroy_team(message)
+      def member_disconnected(message)
+        if player_team = PlayerTeam.find!(message.player_id)
+          leave = LeaveTeam.new(
+            :name => player_team.name,
+            :player_id => message.player_id
+          )
+          leave_team(leave)
+        end
+      end
+
+      def destroy_team(message,force=false)
         if team = Team.find!(message.name)
-          if team.owner == message.player_id
+          if team.owner == message.player_id || force
             team.members.each do |member|
               leave_chat(message.name,member)
               destroy_player_team(member)
@@ -92,14 +159,27 @@ module GameMachine
             end
             commands.datastore.call_dbproc(:team_remove_team,'teams',team.to_storage_entity,false)
             team.destroy!
-            GameMachine.logger.info "Team #{team.name} destroyed"
+            handler_team_removed(team.name)
           end
         end
         send_team_left(message.name,message.player_id)
       end
 
       def join_team(message,invite_accepted=false)
+        unless can_add_member?(message.name,message.player_id)
+          return false
+        end
+
         if team = Team.find!(message.name)
+
+          if (team.max_members.nil? || team.max_members == 0)
+            team.max_members = default_max_members
+          end
+
+          if team.members.size >= team.max_members
+            return
+          end
+
           if team.access == 'public' || invite_accepted
             unless team.members.include?(message.player_id)
               team.members << message.player_id
@@ -115,21 +195,41 @@ module GameMachine
 
       def leave_team(message)
         if team = Team.find!(message.name)
-          if message.player_id == team.owner
+          if message.player_id == team.owner && destroy_on_owner_leave?
             destroy_team(
               DestroyTeam.new(
                 :name => team.name,
                 :player_id => message.player_id
               )
             )
-          else
-            team.members.delete_if {|member| member == message.player_id}
-            team.save!
-            commands.datastore.call_dbproc(:team_update_team,'teams',team.to_storage_entity,false)
-            destroy_player_team(message.player_id)
-            send_team_left(message.name,message.player_id)
-            leave_chat(message.name,message.player_id)
+            return
           end
+
+
+          # Last member, just destroy
+          if team.members.include?(message.player_id) && team.members.size == 1
+            destroy_team(
+              DestroyTeam.new(
+                :name => team.name,
+                :player_id => message.player_id
+              ),
+              true
+            )
+            return
+          end
+
+          team.members.delete_if {|member| member == message.player_id}
+
+          # Reassign owner if they leave
+          if team.owner == message.player_id
+            team.owner = team.members.first
+          end
+
+          team.save!
+          commands.datastore.call_dbproc(:team_update_team,'teams',team.to_storage_entity,false)
+          destroy_player_team(message.player_id)
+          send_team_left(message.name,message.player_id)
+          leave_chat(message.name,message.player_id)
         else
           send_team_left(message.name,message.player_id)
         end
@@ -152,13 +252,21 @@ module GameMachine
         end
       end
 
-      def teams_request(message)
-        if teams = Teams.find('teams')
-          commands.player.send_message(teams,message.player_id)
-        end
-        if player_team = PlayerTeam.find!(message.player_id)
-          if team = Team.find!(player_team.name)
-            commands.player.send_message(team,message.player_id)
+      def find_match(message)
+        if team = Team.find(message.team_name)
+          if match_data = handler_find_match(message.team_name)
+            if opponent = Team.find(match_data.fetch(:team_name))
+              key = team.name + opponent.name
+              server = Akka.instance.hashring.bucket_for(key)
+              match = Match.new(
+                :server => server,
+                :game_handler => match_data.fetch(:game_handler,nil)
+              )
+              participants = team.members + opponent.members
+              participants.each do |participant|
+                commands.player.send_message(match,participant)
+              end
+            end
           end
         end
       end
@@ -166,7 +274,7 @@ module GameMachine
       def on_receive(message)
         if message.is_a?(CreateTeam)
           create_team(message)
-          teams_request(message)
+          send_team(message)
         elsif message.is_a?(DestroyTeam)
           destroy_team(message)
         elsif message.is_a?(TeamInvite)
@@ -175,11 +283,20 @@ module GameMachine
           team_accept_invite(message)
         elsif message.is_a?(JoinTeam)
           join_team(message)
-          teams_request(message)
+          send_team(message)
         elsif message.is_a?(LeaveTeam)
           leave_team(message)
         elsif message.is_a?(TeamsRequest)
           teams_request(message)
+        elsif message.is_a?(FindMatch)
+          find_match(message)
+        elsif message.is_a?(MessageLib::ClientManagerEvent)
+          if message.event == 'disconnected'
+            member_disconnected(message)
+            GameMachine.logger.info "#{self.class.name} #{message.player_id} removed from team(disconnected)"
+          end
+        else
+          unhandled(message)
         end
       end
